+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Bug Fixes
 1. [#173](https://github.com/influxdata/influxdb-client-java/pull/173): Query error could be after _success_ table
 1. [#176](https://github.com/influxdata/influxdb-client-java/pull/176): Blocking API batches Point by precision
+1. [#180](https://github.com/influxdata/influxdb-client-java/pull/180): Fixed concatenation of url
 
 ## 1.13.0 [2020-10-30]